.history/requirements_20251019224040.txt
pygame>=2.0.0
Pillow>=9.0.0
numpy>=1.20.0
pyttsx3>=2.90
